packages/stepproof-runtime/pyproject.toml
[project]
name = "stepproof-runtime"
version = "0.0.1"
description = "StepProof control plane — policy engine, verifier dispatch, audit log"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "BUSL-1.1" }
authors = [{ name = "Eidos AGI", email = "daniel@eidosagi.com" }]

dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "pydantic>=2.9",
    "pyyaml>=6.0",
    "aiosqlite>=0.20",
]

[project.scripts]
stepproof-runtime = "stepproof_runtime.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stepproof_runtime"]
